+++ bill/discounts.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/currency"
 	"github.com/invopop/gobl/i18n"
 	"github.com/invopop/gobl/num"
 	"github.com/invopop/gobl/tax"
@@ -179,7 +180,8 @@ func (Discount) JSONSchemaExtend(schema *jsonschema.Schema) {
 	extendJSONSchemaWithDiscountKey(schema)
 }
 
-func calculateDiscounts(lines []*Discount, sum, zero num.Amount, rr cbc.Key) {
+func calculateDiscounts(lines []*Discount, cur currency.Code, sum num.Amount, rr cbc.Key) {
+	zero := cur.Def().Zero()
 	if len(lines) == 0 {
 		return
 	}
@@ -187,36 +189,23 @@ func calculateDiscounts(lines []*Discount, sum, zero num.Amount, rr cbc.Key) {
 		l.Index = i + 1
 		if l.Percent != nil && !l.Percent.IsZero() {
 			base := sum
-			exp := zero.Exp()
 			if l.Base != nil {
-				base = l.Base.RescaleUp(exp)
-				exp = base.Exp()
+				base = l.Base.RescaleUp(zero.Exp())
 			}
 			l.amount = l.Percent.Of(base)
-			switch rr {
-			case tax.RoundingRuleRoundThenSum:
-				l.amount = l.amount.Rescale(zero.Exp())
-				l.Amount = l.amount
-			default:
-				l.Amount = l.amount.Rescale(exp)
-			}
+			l.amount = tax.ApplyRoundingRule(rr, cur, l.amount)
 		} else {
-			switch rr {
-			case tax.RoundingRuleRoundThenSum:
-				l.Amount = l.Amount.Rescale(zero.Exp())
-			default:
-				l.Amount = l.Amount.MatchPrecision(zero)
-			}
-			l.amount = l.Amount
+			l.amount = l.Amount.Rescale(zero.Exp())
 		}
+		l.Amount = l.amount.Rescale(zero.Exp())
 	}
 }
 
-func calculateDiscountSum(discounts []*Discount, zero num.Amount) *num.Amount {
+func calculateDiscountSum(discounts []*Discount, cur currency.Code) *num.Amount {
 	if len(discounts) == 0 {
 		return nil
 	}
-	total := zero
+	total := cur.Def().Zero()
 	for _, l := range discounts {
 		total = total.MatchPrecision(l.amount)
 		total = total.Add(l.amount)