+++ CHANGELOG.md
@@ -8,7 +8,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed
 
-- `bill`: `round-then-sum` rounding rule _now_ implies that line totals will not be calculated with additional precision.
+- `bill`: line totals will be rounded to currency precision for presentation only
+- `bill`: Discount and Charge amounts always rounded to currency's precision
+- `bill`: `round-then-sum` rounding rule _now_ implies that line totals will not be calculated with additional precision, this brings closer aliance with EN16931 requirements.
 
 ## [v0.211.0] - 2025-02-28